Combine all snippets and required stubs into a single snippets file for each of SE and MP
