+++ docs/src/main/asciidoc/se/openapi/openapi-generator.adoc
@@ -25,9 +25,7 @@
 include::{rootdir}/includes/se.adoc[]
 :incdir: {rootdir}/includes/openapi
 :gen-inc: {incdir}/openapi-generator.adoc
-:petServiceImpl: {sourcedir}/se/openapi/gen/server/api/PetServiceImplCustom.java
-:addPetOpCustom: {sourcedir}/se/openapi/gen/server/api/AddPetOpCustom.java
-:exampleClient: {sourcedir}/se/openapi/gen/client/ExampleClient.java
+:snippets: {sourcedir}/se/openapi/OpenApiGeneratorSnippets.java
 
 :helidon-client-xref: {webclient-page}
 
@@ -57,7 +55,7 @@ The implementation class contains a `handle` method for each API operation with 
 [source,java]
 .The generated `handleAddPet` method in the `PetApiImpl` class
 ----
-include::{petServiceImpl}/[tags=class-header;handleAddPetDecl;handleAddPetGenerated;handleAddPetEnd, indent=0]
+include::{snippets}/[tags=PetServiceImplCustomclass-header;PetServiceImplCustomhandleAddPetDecl;PetServiceImplCustomhandleAddPetGenerated;PetServiceImplCustomhandleAddPetEnd, indent=0]
 }
 ----
 
@@ -65,7 +63,7 @@ Customize the class to manage the pets and revise the method to save the new pet
 [source,java]
 .The customized `handleAddPet` method in the `PetApiImpl` class
 ----
-include::{petServiceImpl}/[tags=class-header;handleAddPetDecl;addedFields;handleAddPetCustom;handleAddPetEnd, indent=0]
+include::{snippets}/[tags=PetServiceImplCustomclass-header;PetServiceImplCustomhandleAddPetDecl;PetServiceImplCustomaddedFields;PetServiceImplCustomhandleAddPetCustom;PetServiceImplCustomhandleAddPetEnd, indent=0]
 }
 ----
 <1> Business logic: create a very simple data store - a real app would use a database.
@@ -79,7 +77,7 @@ The following example illustrates this for the `findPetsByTags` operation and it
 [source,java]
 .The customized `findPetsByTags` method in the `PetApiImpl` class
 ----
-include::{petServiceImpl}/[tags=class-header;addedFields;handleFindPetsByTagsDecl;handleFindPetsByTagsCustom;handleFindPetsByTagsEnd, indent=0]
+include::{snippets}/[tags=PetServiceImplCustomclass-header;PetServiceImplCustomaddedFields;PetServiceImplCustomhandleFindPetsByTagsDecl;PetServiceImplCustomhandleFindPetsByTagsCustom;PetServiceImplCustomhandleFindPetsByTagsEnd, indent=0]
 }
 ----
 <1> Uses the same data store as in the earlier example.
@@ -115,14 +113,14 @@ Override how a parameter is extracted by following these steps, using the `AddPe
 .Customized `AddPetOp` class
 [source,java]
 ----
-include::{addPetOpCustom}[tag=classDecl, indent=0]
+include::{snippets}[tag=AddPetOpCustomClassDecl, indent=0]
 ----
 . In that subclass override the relevant method.
 +
 .Customized `pet` method for extracting and validating the `pet` parameter
 [source,java]
 ----
-include::{addPetOpCustom}[tag=petMethod, indent=0]
+include::{snippets}[tag=AddPetOpCustompetMethod, indent=0]
 ----
 <1> Extracts the parameter from the request. This happens to use the same logic as in the generated method but you can customize that as well if you need to.
 <2> Apply any relevant validations. This silly but illustrative example rejects any pet name that starts with a lower-case letter.
@@ -132,7 +130,7 @@ include::{addPetOpCustom}[tag=petMethod, indent=0]
 [source,java]
 .Providing your custom implementation of `AddPet`
 ----
-include::{petServiceImpl}[tag=createAddPetOp, indent=0]
+include::{snippets}[tag=PetServiceImplCustomcreateAddPetOp, indent=0]
 ----
 
 ===== Override how an operation is prepared from a request
@@ -172,7 +170,7 @@ In the simplest case, your code can get an `ApiClient` instance directly.
 [source,java]
 .Creating an `ApiClient` instance - simple case
 ----
-include::{exampleClient}[tags=classDecl;initDecl;defaultApiClient;initEnd;classEnd, indent=0]
+include::{snippets}[tags=ExampleClientclassDecl;ExampleClientinitDecl;ExampleClientDefaultApiClient;ExampleClientinitEnd;ExampleClientclassEnd, indent=0]
 ----
 <1> The same `ApiClient` instance can be reused to invoke multiple APIs handled by the same server.
 <2> Creates an `ApiClient` instance using default settings from the OpenAPI document.
@@ -189,7 +187,7 @@ Your code does not need to know how the object mapper setting is conveyed to the
 [source,java]
 .Creating an `ApiClient` instance - influencing the `ApiClient.Builder`
 ----
-include::{exampleClient}[tags=classDecl;initDecl;apiClientWithObjMapper;initEnd;classEnd, indent=0]
+include::{snippets}[tags=ExampleClientclassDecl;ExampleClientinitDecl;ExampleClientApiClientWithObjMapper;ExampleClientinitEnd;ExampleClientclassEnd, indent=0]
 ----
 <1> Stores a reusable `ApiClient`.
 <2> A real app would fully set up the `ObjectMapper`.
@@ -201,7 +199,7 @@ In more complicated situations, your code can adjust the settings of the `WebCli
 [source,java]
 .Creating an `ApiClient` instance - adjusting the `WebClientConfig.Builder`
 ----
-include::{exampleClient}[tags=classDecl;initDecl;apiClientWithAdjustedBuilder;initEnd;classEnd, indent=0]
+include::{snippets}[tags=ExampleClientclassDecl;ExampleClientinitDecl;ExampleClientApiClientWithAdjustedBuilder;ExampleClientinitEnd;ExampleClientclassEnd, indent=0]
 ----
 <1> Stores a reusable `AppClient`.
 <2> Creates a new `AppClient` builder.
@@ -217,7 +215,7 @@ Lastly, you can construct the `WebClientConfig.Builder` entirely yourself and ha
 [source,java]
 .Creating an `ApiClient` instance - using a custom `WebClientConfig.Builder`
 ----
-include::{exampleClient}[tags=classDecl;initDecl;apiClientWithCustomWebClientBuilder;initEnd;classEnd, indent=0]
+include::{snippets}[tags=ExampleClientclassDecl;ExampleClientinitDecl;ExampleClientApiClientWithCustomWebClientBuilder;ExampleClientinitEnd;ExampleClientclassEnd, indent=0]
 ----
 <1> Stores a reusable `AppClient`.
 <2> Creates a new `WebClientConfig.Builder`.
@@ -241,7 +239,7 @@ To invoke an operation defined on the `PetApi` interface, your code instantiates
 [source,java]
 .Preparing the PetStore Client API
 ----
-include::{exampleClient}[tags=classDeclFull;prepPetApi;classEnd, indent=0]
+include::{snippets}[tags=ExampleClientclassDeclFull;ExampleClientPrepPetApi;ExampleClientclassEnd, indent=0]
 ----
 <1> Stores a reusable `AppClient`.
 <2> Stores a reusable `PetApi` for invoking pet-related operations.
@@ -277,7 +275,7 @@ You can adopt different styles of retrieving the results, depending on the speci
 [source,java]
 .Access with only result access
 ----
-include::{exampleClient}[tags=findAvailableDecl;findAvailableSimple;findAvailableUseResult;findAvailableEnd, indent=0]
+include::{snippets}[tags=ExampleClientFindAvailableDecl;ExampleClientFindAvailableSimple;ExampleClientFindAvailableUseResult;ExampleClientFindAvailableEnd, indent=0]
 ----
 <1> Use the previously-prepared `petApi` to find pets that have the `available` status.
 <2> Retrieve the typed result from the `ApiResponse`.
@@ -290,7 +288,7 @@ The next example shows how your code can use the `HTTPClientResponse`.
 [source,java]
 .Access with status checking
 ----
-include::{exampleClient}[tags=findAvailableDecl;findAvailableSimple;findAvailableResponseCheck;findAvailableEnd, indent=0]
+include::{snippets}[tags=ExampleClientFindAvailableDecl;ExampleClientFindAvailableSimple;ExampleClientFindAvailableResponseCheck;ExampleClientFindAvailableEnd, indent=0]
 ----
 <1> Start the remote service invocation.
 <2> Wait for the HTTP response status and headers to arrive.
