--- docs/src/main/java/io/helidon/docs/mp/openapi/ExampleOpenApiGenClientResource.java
@@ -1,50 +1,0 @@
-/*
- * Copyright (c) 2024 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.helidon.docs.mp.openapi;
-
-import jakarta.inject.Inject;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.PathParam;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
-
-import org.eclipse.microprofile.rest.client.inject.RestClient;
-
-interface Pet {
-}
-interface PetApi {
-    Pet getPetById(long petId);
-}
-class ApiException extends Exception {
-}
-
-// tag::class-declaration[]
-@Path("/exampleServiceCallingService") // <1>
-public class ExampleOpenApiGenClientResource {
-    @Inject // <2>
-    @RestClient // <3>
-    private PetApi petApi; // <4>
-
-    @GET
-    @Path("/getPet/{petId}")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Pet getPetUsingId(@PathParam("petId") Long petId) throws ApiException {
-        Pet pet = petApi.getPetById(petId); // <5>
-        return pet;
-    }
-}
-// end::class-declaration[]
